[feat] finished delayed packets

+++ simulator_main/dtos/GetErrorSimulationDto.cs
@@ -7,14 +7,14 @@ namespace simulator_main.dtos
 {
     public class GetErrorSimulationDto
     {
-        public GetErrorSimulationDto(string icdName,float packetDelayAmount,int packetNoiseAmount)
+        public GetErrorSimulationDto(string icdName,int packetDelayAmount,int packetNoiseAmount)
         {
             this.IcdName = icdName;
             this.PacketDelayAmount = packetDelayAmount;
             this.PacketNoiseAmount = packetNoiseAmount;
         }
         public string IcdName { get; set; }
-        public float PacketDelayAmount { get; set; }
+        public int PacketDelayAmount { get; set; }
         public int PacketNoiseAmount { get; set; }
     }
 }
